+++ lib/dart_container.dart
@@ -17,5 +17,6 @@ export 'src/scheduler/scheduled_job.dart';
 export 'src/scheduler/scheduled_job_type.dart';
 export 'src/scheduler/scheduler.dart';
 export 'src/container/not_found_handler.dart';
+export 'src/container/abstract_topic_handler.dart';
 
 export 'package:dart_router_extended/dart_router_extended.dart';