+++ lib/src/container/abstract_topic_handler.dart
@@ -1,0 +1,10 @@
+import 'package:dart_container/src/container/container.dart';
+
+abstract class AbstractTopicHandler {
+  TopicHandler buildTopicHandler();
+  TopicHandler? _handler;
+
+  TopicHandler getTopicHandler() {
+    return _handler ??= buildTopicHandler();
+  }
+}